Weitere Kataloge und schuljahresbezogene Bereitstellung der Informationen für Organisationseinheiten

+++ svws-schulen/src/main/java/de/svws_nrw/schulen/v1/data/SchuldateiOrganisationseinheitAdresse.java
@@ -44,7 +44,7 @@ public class SchuldateiOrganisationseinheitAdresse extends SchuldateiEintrag {
 
 	/** Qualität der Verortung */
 	@Schema(description = "Qualität der Verortung")
-	public long qualitaetverortung;
+	public int qualitaetverortung;
 
 	/** Koordinatenrechtswert der Adresse */
 	@Schema(description = "Koordinatenrechtswert der Adresse")
